pyproject.toml
[project]
name = "backend-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-ninja>=1.4.3",
    "ipykernel>=6.30.1",
    "mysqlclient>=2.2.7",
    "pandas>=2.3.3",
    "pymysql>=1.1.2",
]
